requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
whitenoise==6.11.0
